+++ tests/Database/SelectFields/ValidateFieldTests/PrivacyDenied.php
@@ -10,7 +10,7 @@ class PrivacyDenied extends Privacy
     /**
      * @inheritDoc
      */
-    public function validate(array $fieldArgs, $queryContext = null): bool
+    public function validate($root, array $fieldArgs, $queryContext = null): bool
     {
         return false;
     }